Branch HEAD
Fix compiler warning.

Original patch by walter@clusterfs.com.

b=12346
i=adilger
r=scjody

+++ lustre/liblustre/tests/sanity.c
@@ -563,7 +563,7 @@ static int check_file_size(char *file, off_t size)
                 return(1);
         }
         if (statbuf.st_size != size) {
-                printf("size of %s: %ld != %lld\n", file, statbuf.st_size, size);
+                printf("size of %s: %ld != %lld\n", file, statbuf.st_size, (unsigned long long )size);
                 return(-1);
         }
         return 0;
@@ -827,7 +827,7 @@ int t23(char *name)
         ret = lseek(fd, -buf_size + 2, SEEK_CUR);
         if (ret != off) {
                 printf("relative seek error for %d %llu != %llu\n",
-                       -buf_size + 2, ret, off);
+                       -buf_size + 2, ret, (unsigned long long) off);
                 if (ret == -1)
                         perror("relative seek");
                 return -1;
@@ -853,7 +853,7 @@ int t23(char *name)
         off = 2048ULL * 1024 * 1024, SEEK_SET;
         ret = lseek(fd, off, SEEK_SET);
         if (ret != off) {
-                printf("seek 2GB error for %llu != %llu\n", ret, off);
+                printf("seek 2GB error for %llu != %llu\n", ret, (unsigned long long) off);
                 if (ret == -1)
                         perror("seek 2GB");
                 return -1;
@@ -992,7 +992,7 @@ int t50b(char *name)
         for (i = 0; i < sizeof(off_array)/sizeof(loff_t); i++) {
                 offset = off_array[i];
                 printf("16 per xfer(total %d), offset %10lld...\t",
-                        _npages, offset);
+                        _npages, (unsigned long long) offset);
                 if (pages_io(16, offset) != 0)
                         return 1;
         }
